appsearch Windows functions
appsearch lacked some functions that macappseach had which prevented the gui from running
cleanly on windows. Pulled the shared functions into a seperate module and imported
them into each of the platform specific modules
